+++ simple_manga_downloader/modules/manga.py
@@ -17,3 +17,8 @@ class BaseManga:
 
     def __len__(self):
         return len(self.chapters)
+
+    @classmethod
+    def check_if_link_matches(cls, link):
+        """Checks if given url is valid for given module"""
+        return cls.site_re.search(link) is not None